+++ libs/bcg_opengl/guis/bcg_gui_property_selector.cpp
@@ -18,14 +18,14 @@ bool gui_property_selector(viewer_state *state, property_container *container, c
     if (property_names.empty()) return false;
     property_names.emplace_back("");
     int current_entry = property_names.size() - 1;
-    for (int i = 0; i < property_names.size(); ++i) {
+    for (size_t i = 0; i < property_names.size(); ++i) {
         if (property_names[i] == current_property_name) {
             current_entry = i;
 
             break;
         }
     }
-    if(current_entry >= property_names.size()){
+    if(current_entry >= static_cast<int>(property_names.size())){
         return false;
     }
     auto change = draw_combobox(&state->window, label.c_str(), current_entry, property_names);
